rename sandbox_snapshot -> snapshot
